Fix/Update/JavaDoc commit
Fixed some flaws in Java-Convention-use.-
Updated "material"-use in CommandItem.
Added JavaDocs to every class.
Added PlugDev as contributor
